Add files via upload
I am still working on creating options. For now, I have tried to implement a basic system where a player can press e to interact with an NPC. They are given a choice of two dialogue options that will affect the ending of the game. I will continue to revise the code before Thursday.