Dev-Update
Dev was in a really bad spot (largely due to my negligence.) This is an edit of the demo designed to fix those issues. Files are in the correct places, everything works correctly, and the bugs are minor. Once battle is added and minor bugs fixed, this will be pushed to main and all future requests will come from this.