Fixed Item Signals and Implemented Equipment and Battle Items
Fixed the logic for item signals so that it no longer only works when you use a healing item.

Then Implemented global weapons, base damage, and equipment to be able to change those values with the swords item.

Changed Custom Signals to send out Effects not Item names, and fixed the poor logic so that it no longer sends out the actual effect instead of unique labels like <Label#120930174727>

Fixed battle logic to update player health based on global to fix conflicts created by using potions and traps.

Fixed the Use/Discard panel visibility when items are used and ESPECIALLY when they can't be used.