Fixed Equipment Name
Fixed Equipment name and added equipment effects as and item property and made sure that it being null didn't mess anything up.

Allowed for more implimentation of equipment.

Tried to change how Usage Panel visibility works, but nothing changed in practice. I would like it if clicking on an item while the panel is up would close it because the overlapping details panel is ugly.