END SCREEN AND (FINAL?) HOTFIX
Hotfix to correct usage of signals. Signals were not being disconnected when battle ended, so keys were unusable after battle.

Same for the inventory UI, but using .Connect and explicitly making it Callable fixes this by disconnected the signals on the QueueFree().

END SCREEN DEMO IN.
Open the door at the bottom of the map, and the end screen loads.
Text based on Karma levels (Neutral, Bad, Worse).